% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CarryingK_vec}
\alias{CarryingK_vec}
\title{Carrying Capacities}
\format{
\subsection{\code{CarryingK_vec}}{

A vector with 10 values (all values = 200).
}
}
\usage{
CarryingK_vec
}
\description{
A vector of carrying capacities.
}
\keyword{datasets}
