% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Env_mat}
\alias{Env_mat}
\title{Environmental Matrix}
\format{
\subsection{\code{Env_mat}}{

Environmental matrix of 1e3 rows and columns with environmental range 0-10 in either dimension
}
}
\usage{
Env_mat
}
\description{
An environmental matrix used for demographic simulations.
}
\keyword{datasets}
