% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Inferred_igraph}
\alias{Inferred_igraph}
\title{Inferred Network}
\format{
\subsection{\code{Inferred_igraph}}{

igraph object containing 10 nodes and 23 links.
}
}
\usage{
Inferred_igraph
}
\description{
Random network treated as though it was an inferred network for package documentation purposes.
}
\keyword{datasets}
