% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Initialise_df}
\alias{Initialise_df}
\title{Initialising Information}
\format{
\subsection{\code{Initialise_df}}{

Data frame containing 400 initialising individuals.
}
}
\usage{
Initialise_df
}
\description{
A data frame object containing initialising individuals.
}
\keyword{datasets}
