% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Initialise_ls}
\alias{Initialise_ls}
\title{Initialising Information}
\format{
\subsection{\code{Initialise_ls}}{

List of 2. (1) data frame containing 400 initialising individuals, (2) containing 10 values of biolcimatic niche preferences.
}
}
\usage{
Initialise_ls
}
\description{
A list object containing data frame of initialising individuals and their bioclimatic niche preferences.
}
\keyword{datasets}
