% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Network_igraph}
\alias{Network_igraph}
\title{Simulated Network}
\format{
\subsection{\code{Network_igraph}}{

igraph object containing 10 nodes and 23 links.
}
}
\usage{
Network_igraph
}
\description{
Random network used to inform simulation framework.
}
\keyword{datasets}
