% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Niches_vec}
\alias{Niches_vec}
\title{Bioclimatic Niche Preferences}
\format{
\subsection{\code{Niches_vec}}{

A vector with 10 values.
}
}
\usage{
Niches_vec
}
\description{
A vector of bioclimatic niche preferences.
}
\keyword{datasets}
