% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rFlatTopNorm.R
\name{rFlatTopNorm}
\alias{rFlatTopNorm}
\title{Random Generation from a Bivariate Flat-Top Normal Distribution}
\usage{
rFlatTopNorm(n, mean = c(0, -1), sd = 1, range = 2, truncDist = 5)
}
\arguments{
\item{n}{A numerical value defining the number of observations to be sampled.}

\item{mean}{A vector of 2 values defining the mean (centre) of the distribution.}

\item{sd}{A numerical value defining the standard deviation of the distribution.}

\item{range}{A numerical value defining the radius of a circle in which the flat-top of the kernel is defined.}

\item{truncDist}{A numerical value defining the distance beyond which values are considered too far from the center of the distribution to be considered reasonable to consider.}
}
\value{
A two-column matrix with the sample from the distribution
}
\description{
This function generates samples from a bivariate flat-top normal distribution.
}
\details{
Rejection sampling is used to carry out the sampling using as a starting point a random sample from a uniform distribution bounded by the truncation distance around the mean. Because the distribution function is continuous and unbounded, conceptually, the truncation could be +/- infinite in both dimensions, but since rejection-sampling is used this would not be computationally efficient. In this respect, we recommend to choose the truncation distance (\code{truncDist}) using a reasonable value for the sampling to be both computationally efficient and representative (so, not too large but not too small).

Even if the function samples from a bivariate flat-top normal distribution, it is designed to focus only on a special case where the standard deviation (\code{sd}) and the radius of a circle in which the flat-top of the kernel is defined (\code{range}) is unique. In doing so, the function assumes weights are equal in both direction.
}
\examples{
rFlatTopNorm(10)
rFlatTopNorm(5, mean = c(10, 12), sd = 3, range = 1, truncDist = 10)

}
\author{
F. Guillaume Blanchet, Département de biologie, mathématiques et des sciences de la santé communautaire, Université de Sherbrooke, Canada.
}
