% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim.Compute.R
\name{Sim.d0Update}
\alias{Sim.d0Update}
\title{Update Dynamic Death Rates}
\usage{
Sim.d0Update(
  ID_df = ID_df,
  which = "Initial",
  event = NULL,
  env.xy = env.xy,
  d0 = d0,
  b0 = b0,
  sd = sd,
  Effect_Mat = Effect_Mat,
  k_vec = k_vec,
  Effect_Dis = Effect_Dis,
  seed = seed
)
}
\arguments{
\item{ID_df}{Data frame of initialising individuals with columns ID, Trait, X, Y, and Species. Output of Sim.Initialise()$ID_df.}

\item{which}{Either Character or Numeric. If Character and == "Initial" then dynbamic death rate is computed for all individuals. If numeric, dynamic death rate is only computed for indviduals affected by the event applied to the individual in the indexed row.}

\item{event}{Character. Either "Birth" or "Death".}

\item{env.xy}{Function with arguments X and Y. Translates X and Y coordinates into optimal local phenotype.}

\item{d0}{Numeric. background death rate (gets altered by the environment and interactions).}

\item{b0}{Numeric. background birth rate (remains constant).}

\item{sd}{Numeric. Habitat suitability in death rate function. Higher values allow individuals to persist in areas of greater environmental maladaptation.}

\item{Effect_Mat}{Weighted adjacency matrix of association/interaction network.}

\item{k_vec}{Named vector containing carrying capacity for each species. Output of Sim.CarryingK().}

\item{Effect_Dis}{Distance to which association/interaction effects are computed around individuals.}

\item{seed}{Numeric. Seed for random processes.}
}
\value{
An updated data frame of individuals and their dynamic death rates.
}
\description{
This is a helper function for updating of dynamic death rates in the simulation run
}
