% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim.Niche.R
\name{Sim.Niche}
\alias{Sim.Niche}
\title{Generate Random Carrying Capacities per Species}
\usage{
Sim.Niche(n_spec = 20, Env_range = c(0, 10), seed = 42)
}
\arguments{
\item{n_spec}{Numeric. Number of species / size of node set.}

\item{Env_range}{Numeric vector of length 2. Minimum and maximum values of environmental gradient.}

\item{seed}{Numeric. Seed for random processes.}
}
\value{
A named vector of random carrying capacity values for each species.
}
\description{
This function generates a vector of random numbers representing carrying capacities for the demographic simulation.
}
\examples{
Sim.Niche(n_spec = 20, Env_range = c(0,10), seed = 42)
Sim.Niche(n_spec = 15, Env_range = c(11,33), seed = 42)

}
\author{
Erik Kusch, Natural History Museum, University of Oslo, Norway.
}
