% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim.Space.r
\name{Sim.Space}
\alias{Sim.Space}
\title{Generate Spatial Environment as Matrix}
\usage{
Sim.Space(
  x_range,
  y_range,
  ncol = 100,
  nrow = 100,
  x_gradient = function(x) x,
  y_gradient = function(y) y
)
}
\arguments{
\item{x_range}{Numeric vector of length 2 specifying the minimum and maximum x-coordinates}

\item{y_range}{Numeric vector of length 2 specifying the minimum and maximum y-coordinates}

\item{ncol}{Integer specifying the number of columns (resolution in x direction). Default is 100.}

\item{nrow}{Integer specifying the number of rows (resolution in y direction). Default is 100.}

\item{x_gradient}{Function that defines how values change along the x-axis. Default is the identity function resulting in a linear gradient.}

\item{y_gradient}{Function that defines how values change along the y-axis. Default is the identity function resulting in a linear gradient.}
}
\value{
A matrix where cells report environmental values while column and row names report X and Y coordinates, respectively.
}
\description{
This function generates a matrix that serves as a space-definition for demographic simulations.
}
\examples{
# 1. Simple linear gradient matrix
mat <- Sim.Space(
    x_range = c(0, 10), y_range = c(0, 10),
    ncol = 1e4, nrow = 1e4,
    x_gradient = function(x) x,
    y_gradient = function(y) y
)

# 2. Homogenuous matrix
mat <- Sim.Space(
    x_range = c(0, 10), y_range = c(0, 10),
    ncol = 1e4, nrow = 1e4,
    x_gradient = function(x) 1,
    y_gradient = function(y) 1
)
}
\author{
Erik Kusch, Natural History Museum, University of Oslo, Norway.
}
