% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SimulationOutput}
\alias{SimulationOutput}
\title{Simulation Output}
\format{
\subsection{\code{SimulationOutput}}{

List of 99, each containing a dataframe of individuals alive at their respective simulation timestep.
}
}
\usage{
SimulationOutput
}
\description{
Output of simulation framework.
}
\keyword{datasets}
