% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Val.Behaviour.R
\name{Val.Behaviour}
\alias{Val.Behaviour}
\title{Network Inference Behaviour Validation}
\usage{
Val.Behaviour(
  model_df,
  mode = "Detection",
  nWarmup = 3000,
  nSamples = 10000,
  nChains = 4,
  nCores = 4,
  seed = 42
)
}
\arguments{
\item{model_df}{A data frame tallying comparisons of inferred and underlying true networks. Rows index pairwise links. Consists of the following columns:
\itemize{
\item Magnitude ... Numeric. Absolute strength/magnitude of link in the true network.
\item SignTrue ... Numeric - -1, 0, or 1. Sign of link in true network. 0 denotes absence of a link.
\item SignInferred ... Numeric - -1, 0, or 1. Sign of link in inferred network. 0 denotes absence of a link.
\item EnvDiff ... Numeric. Difference in environmental niche preference of linked species.
See examples for a workflow of how to create this data frame.
}}

\item{mode}{Character. Either "Inference" or "Detection". Mode of behaviour validation:
\itemize{
\item Inference ... likelihood of assignment of a link irrespective of whether it is correctly assigned
\item Detection ... likelihood of correct inference of individual links
}}

\item{nWarmup}{Integer. Number of warmup iterations for Bayesian models.}

\item{nSamples}{Integer. Number of total iterations for Bayesian models.}

\item{nChains}{Integer. Number of markov chains to run.}

\item{nCores}{Integer. Number of cores to utilise.}

\item{seed}{Numeric. Seed for random processes.}
}
\value{
A list with three elements:
\itemize{
\item Positive ... brmsfit object describing the Bayesian model run to identify either inference or detection behaviour for positive inferred links
\item Negative ... brmsfit object describing the Bayesian model run to identify either inference or detection behaviour for negative inferred links
\item Absent ... brmsfit object describing the Bayesian model run to identify either inference or detection behaviour for absent inferred links
}
}
\description{
Logistic Bayesian models of network inference approach behaviour. Runs models per link sign (positive, negative, absent) to identify the effects of underlying link magnitude and environmental niche preference differences.
}
\examples{
## loading data ---
data("Network_igraph")
data("Inferred_igraph")
data("Niches_vec")
## Reformatting data ---
### Make igraphs into matrices
matrices_ls <- lapply(list(Network_igraph, Inferred_igraph), FUN = function(NetworkI){
  if(is.null(igraph::V(NetworkI)$names)){
    igraph::V(NetworkI)$names <- paste0("Sp_", igraph::V(NetworkI))
  }
  mat <- as.matrix(igraph::as_adjacency_matrix(NetworkI, attr = "weight"))
  colnames(mat) <- rownames(mat) <- igraph::V(NetworkI)$names
  diag(mat) <- NA
  if(!igraph::is_directed(NetworkI)){
    mat[lower.tri(mat) ] <- NA
  }
  mat
})

### Calculate environmental differences
SPTrait_df <- data.frame(Niches_vec)
SPTrait_df$Species <- rownames(SPTrait_df)
colnames(SPTrait_df) <- c("Trait", "Species")
SPTrait_mat <- abs(outer(SPTrait_df$Trait, SPTrait_df$Trait, '-'))
colnames(SPTrait_mat) <- rownames(SPTrait_mat) <- colnames(SPTrait_df$Species)

## Collate data frame for behaviour analysis ---
model_df <- data.frame(
  Magnitude = abs(as.vector(matrices_ls[[1]])),
  SignTrue = as.vector(sign(matrices_ls[[1]])),
  SignInferred = as.vector(sign(matrices_ls[[2]])),
  EnvDiff = as.vector(SPTrait_mat)
)
model_df <- na.omit(model_df)
\dontrun{
(Detection <- Val.Behaviour(model_df, mode = "Detection"))
(Inference <- Val.Behaviour(model_df, mode = "Inference"))
}

}
