% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Val.ErrorRates.R
\name{Val.ErrorRates}
\alias{Val.ErrorRates}
\title{Network Inference Error Rates Validation}
\usage{
Val.ErrorRates(Network1, Network2, NetworkType = "Association")
}
\arguments{
\item{Network1}{matrix object. First network in comparison.}

\item{Network2}{matrix object. Second network in comparison.}

\item{NetworkType}{Character. Whether to create an interaction or association matrix. In an interaction matrix, columns act on rows. In an association matrix, entries are mirrored around the diagonal. Possible values: "Association" or "Interaction".}
}
\value{
A dataframe containing percentages and error rate types:
\itemize{
\item TP ... true positive associations; how many of the inferred positive associations are correctly identified as such? - NaN means no positives in the inferred matrix
\item TN ... true negative associations; how many of the inferred negative associations are correctly identified as such? - NaN means no negatives in the inferred matrix
\item TA ... true absent; how many of the inferred absent interactions are correctly identified as such? - NaN means no absents in the inferred matrix
\item FP ... falsely inferred positive; how many of the inferred positive associations are incorrectly identified as such?
\item FN ... falsely inferred negative; how many of the inferred negative associations are incorrectly identified as such?
\item FA ... false absent; how many of the inferred absent interactions are correctly identified as such?
\item MP ... true positive links which were not inferred; how many of the true positive associations were not inferred?
\item MN ... true negative links which were not inferred; how many of the true negative associations were not inferred?
\item MA ... true absent links which were not inferred; how many of the true absent associations were not inferred?
}
}
\description{
Calculate accuracy of network inference by sign of underlying true link.
}
\examples{
data("Network_igraph")
data("Inferred_igraph")
Val.ErrorRates(Network1 = Network_igraph, Network2 = Inferred_igraph)

}
\author{
Erik Kusch, Natural History Museum, University of Oslo, Norway.
}
