% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Val.Realise.R
\name{Val.Realise}
\alias{Val.Realise}
\title{Reduction of Network to Realisable Component}
\usage{
Val.Realise(Network_igraph, Trait_means, Effect_Dis = 0.5, Env_sd = 2.5)
}
\arguments{
\item{Network_igraph}{An igraph object with association/interaction strength stored as "weight" attribute of edges. Output of Sim.Network().}

\item{Trait_means}{Numeric. Mean for bioclimatic niche preference of each species.}

\item{Effect_Dis}{Numeric. Distance within which neighbouring individuals interact with a focal individual.}

\item{Env_sd}{Numeric. Habitat suitability in death rate function. Higher values allow individuals to persist in areas of greater environmental maladaptation.}
}
\value{
An igraph object with association/interaction strength stored as "weight" attribute of edges.
}
\description{
Takes an association or interaction network object as well as the bioclimatic niche preferences of each species to compute which links stored in the network can be realised. The cutoff is calculated as the difference between the niche preferences of species pairs vs. Effect_Dis + Env_sd.
}
\examples{
data("Niches_vec")
data("Network_igraph")
Val.Realise(Network_igraph = Network_igraph, Niches_vec = Niches_vec)

}
