% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim.Compute.R
\name{SIM.Compute}
\alias{SIM.Compute}
\title{Carry out Demographic Simulation}
\usage{
SIM.Compute(
  d0 = 0.4,
  b0 = 0.6,
  env.xy = function(x = NULL, y = NULL) {
     x
 },
  t_max = 10,
  t_inter = 0.1,
  sd = 2.5,
  migration = 0.2,
  Effect_Dis = 0.5,
  Network_igraph,
  k_vec,
  ID_df,
  Env_range,
  seed = 42,
  verbose = TRUE,
  RunName = ""
)
}
\arguments{
\item{d0}{Numeric. background death rate (gets altered by the environment and interactions).}

\item{b0}{Numeric. background birth rate (remains constant).}

\item{env.xy}{Function with arguments X and Y. Translates X and Y coordinates into optimal local phenotype.}

\item{t_max}{Numeric. Maximum simulation time.}

\item{t_inter}{Numeric. Interval length at which to record simulation outputs (measured in simulation time).}

\item{sd}{Numeric. Habitat suitability in death rate function. Higher values allow individuals to persist in areas of greater environmental maladaptation.}

\item{migration}{Numeric. Standard deviation of 0-centred normal dsitribution from which natal dispersal is drawn.}

\item{Effect_Dis}{Numeric. Distance within which neighbouring individuals interact with a focal individual.}

\item{Network_igraph}{An igraph object with association/interaction strength stored as "weight" attribute of edges. Output of Sim.Network().}

\item{k_vec}{Named vector containing carrying capacity for each species. Output of Sim.CarryingK().}

\item{ID_df}{Data frame of initialising individuals with columns ID, Trait, X, Y, and Species. Output of Sim.Initialise()$ID_df.}

\item{Env_range}{Numeric vector of length 2. Minimum and maximum values of environmental gradient.}

\item{seed}{Numeric. Seed for random processes.}

\item{verbose}{Logical. Whether to print simulation time and sampling interval.}

\item{RunName}{Character. Name for .RData object written to disk.}
}
\value{
A list containing data frame object with the same columns as ID_df at each sampling interval defined via n_inter until t_max is reached.
}
\description{
This function carries out spatially explicit, individual-based demographic simulation for the generation of data products ready for network inference.
}
\examples{
Network_igraph <- SIM.Network(n_spec = 10, NetworkType = "Association", Sparcity = 0.5, MaxStrength = 1, seed = 42)
CarryingK_vec <- Sim.CarryingK(n_spec = 10, k_range = c(200,200), seed = 42)
Initialise_ls <- Sim.Initialise(n_spec = 10, n_individuals = 4e2,n_mode = "total", Env_range = c(0, 10), Trait_sd = 1, seed = 42)
SIM.Compute(d0 = 0.4,
            b0 = 0.6,
            env.xy = function(x = NULL, y = NULL){x},
            t_max = 10,
            t_inter = 0.1,
            sd = 2.5,
            migration = 0.2,
            Effect_Dis = 0.5,
            Network_igraph,
            k_vec = CarryingK_vec, 
            ID_df = Initialise_ls$ID_df,
            Env_range = c(0, 10),
            seed = 42,
            verbose = TRUE, # whether to print progress in time as current time
            RunName = "TestRun"
)

}
