% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim.Initialise.R
\name{Sim.Initialise}
\alias{Sim.Initialise}
\title{Generate Initialising Individuals for Demographic Simulation}
\usage{
Sim.Initialise(
  n_spec = 10,
  n_individuals = 400,
  n_mode = "each",
  Env_range = c(0, 10),
  Trait_means,
  Trait_sd = 1,
  seed = 42
)
}
\arguments{
\item{n_spec}{Numeric. Number of species / size of node set.}

\item{n_individuals}{Numeric. How many initialising individuals to randomly create.}

\item{n_mode}{Character. How to use n_individuals. If "each" the the number of initialising individuals will be equal to n_individual multiplied by n_spec. If "total", n_individuals are created with random species memberships within n_spec.}

\item{Env_range}{Numeric vector of length 2. Minimum and maximum values of environmental gradient.}

\item{Trait_means}{Numeric. Mean for bioclimatic niche preference of each species. Must be same length as number specified in n_spec.}

\item{Trait_sd}{Numeric. Standard deviation of trait distributions from which trait values of each individual are drawn.}

\item{seed}{Numeric. Seed for random processes.}
}
\value{
A data frame - the data frame of initialising individuals with columns ID, Trait, X, Y, and Species.
}
\description{
This function generates a set of individuals placed at random across the simulation landscape in x and y dimensions. It also applies species specific niche preferences to each individual with respect to their species membership.
}
\examples{
Niches_vec <- Sim.Niche(n_spec = 10, Env_range = c(0, 10), seed = 42)
Sim.Initialise(n_spec = 10, n_individuals = 4e2,n_mode = "each",
               Env_range = c(0, 10), Trait_means = Niches_vec, Trait_sd = 1, seed = 42)
Sim.Initialise(n_spec = 20, n_individuals = 4e2,n_mode = "total",
               Env_range = c(10, 100), Trait_means = Niches_vec, Trait_sd = 2.5, seed = 42)

}
